Adds CharacterHarm and harm field to Character, setCharacterHarm mutation
