+++ awc-web/src/main/resources/schema.graphqls
@@ -225,6 +225,16 @@ type HxStat {
     hxValue: Int!
 }
 
+type CharacterHarm {
+    id: ID!
+    value: Int!
+    isStabilized: Boolean!
+    hasComeBackHard: Boolean!
+    hasComeBackWeird: Boolean!
+    hasChangedPlaybook: Boolean!
+    hasDied: Boolean!
+}
+
 type Character {
     id: ID!
     name: String
@@ -237,6 +247,7 @@ type Character {
     playbookUnique: PlaybookUnique
     characterMoves: [CharacterMove]
     barter: Int
+    harm: CharacterHarm
 }
 
 type GameRole {
@@ -314,12 +325,22 @@ type StatModifier {
     modification: Int!
 }
 
-# ---------------------------------------------- QUERIES --------------------------------------------- #
+# ---------------------------------------------- INPUTS --------------------------------------------- #
 
 input HxInput {
     characterId: String!
     characterName: String!
     hxValue: Int!
+}
+
+input HarmInput {
+    id: ID!
+    value: Int!
+    isStabilized: Boolean!
+    hasComeBackHard: Boolean!
+    hasComeBackWeird: Boolean!
+    hasChangedPlaybook: Boolean!
+    hasDied: Boolean!
 }
 # ---------------------------------------------- QUERIES --------------------------------------------- #
 
@@ -388,6 +409,8 @@ type Mutation {
         hxStats: [HxInput]!): Character
     adjustCharacterHx(gameRoleId: String!,
         characterId: String!, hxId: String!, value: Int!): Character
+    setCharacterHarm(gameRoleId: String!,
+        characterId: String!, harm: HarmInput!): Character
     finishCharacterCreation(gameRoleId: String!,
         characterId: String!): Character
     finishPreGame(gameId: String!): Game
